+++ src/main/java/shop/dongho/repository/ProducerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import shop.dongho.model.Producer;
 
-public interface ProducerRepository extends PagingAndSortingRepository<Producer, Long> {
+public interface ProducerRepository extends PagingAndSortingRepository<Producer, Integer> {
     Page<Producer> findByNameContaining(String s, Pageable pageable);
 }
