+++ src/main/java/shop/dongho/repository/ProductRepository.java
@@ -8,12 +8,12 @@ import shop.dongho.model.Product;
 public interface ProductRepository extends PagingAndSortingRepository<Product, Integer> {
     Page<Product> findAllByNameContaining(String s, Pageable pageable);
 
-    Page<Product> findAllByProducer_Id(Long id,Pageable pageable);
+    Page<Product> findAllByProducer_Id(Integer id,Pageable pageable);
 
     Page<Product> findAllByUnitPriceLessThan(Integer number, Pageable pageable);
 
     Page<Product> findAllByUnitPriceBetween(Integer pri3,Integer pri5,Pageable pageable);
 
-    Page<Product> findAllByProductType_Id(Long id, Pageable pageable);
+    Page<Product> findAllByProductType_Id(Integer id, Pageable pageable);
 
 }
